Move locale to base Styler; use locale in PieStyler number formatting
- Move locale field, getLocale(), and setLocale() from AxesChartStyler
  up to the base Styler class so all chart types (including Pie) can
  benefit from locale-aware number formatting
- PlotContent_Pie now constructs DecimalFormat with DecimalFormatSymbols
  derived from the styler locale, fixing locale-dependent decimal
  separators in pie slice labels and tooltips
- Remove stale TODO comment from ChartButton

Fixes #868
